Introduce Vue framework
