+++ src/main.ts
@@ -1,0 +1,10 @@
+import './assets/base.css'
+import { createApp } from 'vue'
+import App from './App.vue'
+import router from './router'
+
+const app = createApp(App)
+
+app.use(router)
+
+app.mount('#app')
